SM removal part 4: delete remaining tests that depend on security manager
